Fixed paycheck system.
Changed datatypes & changed a wrong if case.

+++ scripts/3_Game/KR_BankingConfigManager.c
@@ -118,7 +118,7 @@ class KR_BankingConfigManager
         return settings;
     }
 
-    float GetCorrectPayCheckTime()
+    int GetCorrectPayCheckTime()
     {
         return PayCheckTickTime * 60000;
     }
